apps/analysis/requirements.txt
numpy
pandas
scipy
hdbscan
sqlalchemy
psycopg2-binary
python-dotenv  # Correct package for managing .env files
pynndescent
scikit-learn